Update to Electron 13
There are many breaking changes between the previous version of Electron
we were using and version 13. The most notable of these is that the
remote module has been deprecated in favor of the ipcRenderer and
ipcMain modules. All code depending on the remote module has been
rewritten to use the new modules. Additionally, contextIsolation has
been enabled by default, so we disable it. This is a temporary measure
until I can figure out what needs to change to make contextIsolation
work. Finally, allowRendererProcessReuse has been enabled by default,
which disables loading non-context-aware modules in the renderer. We
disable this as well so that we can load Discord.js in the renderer.
This is also temporary.