Update to Electron 16 and Discord.js 13
Turns out the context-aware stuff only applies to native modules, which
Discord.js is not. So we can just update to the latest version of
Discord.js and Electron and everything works fine.

Currently, getting the list of channels and switching to the proper one
is broken. This will be fixed in the next commit.